src/csv-gen-http/requirements.txt
functions-framework==3.*
google-cloud-storage==2.3.*
google-cloud-bigquery==3.20.*
google-cloud-logging==3.10.*
flask==2.*
pandas==2.*
db-dtypes==1.*